fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int fib[50];
+	unsigned long long int fib[50];
 	int i;
 
 	fib[0] = 1;
@@ -18,8 +18,8 @@ int main(void)
 	}
 	for (i = 0; i < 49; i++)
 	{
-		printf("%d, ", fib[i]);
+		printf("%llu, ", fib[i]);
 	}
-	printf("%d\n", fib[49]);
+	printf("%llu\n", fib[49]);
 	return (0);
 }
